Chore: seprate test into linux env which can't be execute in darwin (#3823)
