+++ pkg/controller/utils/capability_test.go
@@ -18,20 +18,15 @@
 package utils
 
 import (
-	"context"
-	"errors"
 	"fmt"
-	"io/ioutil"
 	"os"
 	"path/filepath"
 	"strings"
 	"testing"
 
-	. "github.com/agiledragon/gomonkey/v2"
 	"github.com/crossplane/crossplane-runtime/pkg/test"
 	"github.com/getkin/kin-openapi/openapi3"
 	"github.com/google/go-cmp/cmp"
-	git "gopkg.in/src-d/go-git.v4"
 	"gotest.tools/assert"
 
 	"github.com/oam-dev/kubevela/apis/core.oam.dev/common"
@@ -422,138 +417,6 @@ variable "bbb" {
 			if tc.want.err == nil {
 				data := string(schema)
 				assert.Equal(t, strings.Contains(data, tc.want.subStr), true)
-			}
-		})
-	}
-}
-
-func TestGetTerraformConfigurationFromRemote(t *testing.T) {
-	// If you hit a panic on macOS as below, please fix it by referencing https://github.com/eisenxp/macos-golink-wrapper.
-	// panic: permission denied [recovered]
-	//    panic: permission denied
-	type want struct {
-		config string
-		errMsg string
-	}
-
-	type args struct {
-		name         string
-		url          string
-		path         string
-		data         []byte
-		variableFile string
-		// mockWorkingPath will create `/tmp/terraform`
-		mockWorkingPath bool
-	}
-	cases := map[string]struct {
-		args args
-		want want
-	}{
-		"valid": {
-			args: args{
-				name: "valid",
-				url:  "https://github.com/kubevela-contrib/terraform-modules.git",
-				path: "",
-				data: []byte(`
-variable "aaa" {
-	type = list(object({
-		type = string
-		sourceArn = string
-		config = string
-	}))
-	default = []
-}`),
-				variableFile: "main.tf",
-			},
-			want: want{
-				config: `
-variable "aaa" {
-	type = list(object({
-		type = string
-		sourceArn = string
-		config = string
-	}))
-	default = []
-}`,
-			},
-		},
-		"configuration is remote with path": {
-			args: args{
-				name: "aws-subnet",
-				url:  "https://github.com/kubevela-contrib/terraform-modules.git",
-				path: "aws/subnet",
-				data: []byte(`
-variable "aaa" {
-	type = list(object({
-		type = string
-		sourceArn = string
-		config = string
-	}))
-	default = []
-}`),
-				variableFile: "variables.tf",
-			},
-			want: want{
-				config: `
-variable "aaa" {
-	type = list(object({
-		type = string
-		sourceArn = string
-		config = string
-	}))
-	default = []
-}`,
-			},
-		},
-		"working path exists": {
-			args: args{
-				variableFile:    "main.tf",
-				mockWorkingPath: true,
-			},
-			want: want{
-				errMsg: "failed to remove the directory",
-			},
-		},
-	}
-
-	for name, tc := range cases {
-		t.Run(name, func(t *testing.T) {
-			if tc.args.mockWorkingPath {
-				err := os.MkdirAll("./tmp/terraform", 0755)
-				assert.NilError(t, err)
-				defer os.RemoveAll("./tmp/terraform")
-				patch1 := ApplyFunc(os.Remove, func(_ string) error {
-					return errors.New("failed")
-				})
-				defer patch1.Reset()
-				patch2 := ApplyFunc(os.Open, func(_ string) (*os.File, error) {
-					return nil, errors.New("failed")
-				})
-				defer patch2.Reset()
-			}
-
-			patch := ApplyFunc(git.PlainCloneContext, func(ctx context.Context, path string, isBare bool, o *git.CloneOptions) (*git.Repository, error) {
-				var tmpPath string
-				if tc.args.path != "" {
-					tmpPath = filepath.Join("./tmp/terraform", tc.args.name, tc.args.path)
-				} else {
-					tmpPath = filepath.Join("./tmp/terraform", tc.args.name)
-				}
-				err := os.MkdirAll(tmpPath, os.ModePerm)
-				assert.NilError(t, err)
-				err = ioutil.WriteFile(filepath.Clean(filepath.Join(tmpPath, tc.args.variableFile)), tc.args.data, 0644)
-				assert.NilError(t, err)
-				return nil, nil
-			})
-			defer patch.Reset()
-
-			conf, err := GetTerraformConfigurationFromRemote(tc.args.name, tc.args.url, tc.args.path)
-			if tc.want.errMsg != "" {
-				if err != nil && !strings.Contains(err.Error(), tc.want.errMsg) {
-					t.Errorf("\n%s\nGetTerraformConfigurationFromRemote(...): -want error %v, +got error:%s", name, err, tc.want.errMsg)
-				}
-			} else {
-				assert.Equal(t, tc.want.config, conf)
 			}
 		})
 	}
